alembic/versions/600cd643061d_nothing_s_done.py
"""nothing's done

Revision ID: 600cd643061d
Revises: 
Create Date: 2025-04-17 23:43:55.685398

"""
from typing import Sequence, Union

from alembic import op
import sqlalchemy as sa


# revision identifiers, used by Alembic.
revision: str = '600cd643061d'
down_revision: Union[str, None] = None
branch_labels: Union[str, Sequence[str], None] = None
depends_on: Union[str, Sequence[str], None] = None


def upgrade() -> None:
    """Upgrade schema."""
    pass


def downgrade() -> None:
    """Downgrade schema."""
    pass
