requirements.txt
flask == 3.1.0
flask-sqlalchemy == 3.1.1
pyodbc == 5.2.0 # For mssql
alembic == 1.15.2